flow_meter: ipfix: fixed timestamp computation
Casting on a bad place caused multiplication on a short data type (32b
instead of 64b).

+++ flow_meter/ipfix-elements.h
@@ -6,8 +6,8 @@
 
 #define BYTES(F)                      F(0,        1,    8,   &flow.octet_total_length)
 #define PACKETS(F)                    F(0,        2,    8,   &flow.pkt_total_cnt)
-#define FLOW_START_MSEC(F)            F(0,      152,    8,   (temp = (uint64_t)(flow.time_first.tv_sec * 1000) + (uint64_t)(flow.time_first.tv_usec / 1000), &temp))
-#define FLOW_END_MSEC(F)              F(0,      153,    8,   (temp = (uint64_t)(flow.time_last.tv_sec * 1000) + (uint64_t)(flow.time_last.tv_usec / 1000), &temp))
+#define FLOW_START_MSEC(F)            F(0,      152,    8,   (temp = ((uint64_t) flow.time_first.tv_sec) * 1000 + (flow.time_first.tv_usec / 1000), &temp))
+#define FLOW_END_MSEC(F)              F(0,      153,    8,   (temp = ((uint64_t) flow.time_last.tv_sec) * 1000 + (flow.time_last.tv_usec / 1000), &temp))
 #define OBSERVATION_MSEC(F)           F(0,      323,    8,   NULL)
 #define INPUT_INTERFACE(F)            F(0,       10,    2,   &this->dir_bit_field)
 #define OUTPUT_INTERFACE(F)           F(0,       14,    2,   NULL)
